Battle UI and Battle Mechanics
_BattleScene battle UI is up and running. In order for it to work
properly when playing you must make a new game from the _MainScene (or
you can start with _EnterPlayerNameScene), and either add party members
through the character menu in the pause screen (press esc key to pause)
OR add enemy characters using the buttons.  The battle does not work
properly if it is only the player character.

Notes:
Yes there is one UI menu that is blank (this is for the player to select
who to effect), it is being worked on
Yes, nothing is actually being effected yet, battle mechanics are a work
in progress.

In Game.cs:
Character boss is null UNLESS it is a boss fight, then set to respective
boss (boss = unicorn)
[] party should be updated when a new party member is added
[] enemyParty should have 1-4 enemy Characters added before a fight and
be reset to null (like in Game.cs) after the fight is over, IF it is a
boss fight with NO additional enemy Characters then leave all null

WARNING:
Do NOT try to make any changes to Battle_Mechanics.c yet, I am still
working on the actual mechanics and may overwrite any work. (and it may
explode...)